Organize code tweaks
