+++ src/node-web-compat.ts
@@ -1,8 +1,12 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 //
-// Compatibility layer, to make this lib work in both Node.js and Web
-// This file just contains stubs for typings
+// To make this library work in both Node.js and Web, some functions require specific code,
+// e.g. in Node.js we can use the "crypto" module, whereas in Web we need to use SubtleCrypto.
+// This file contains an interface that the specific Node.js and Web implementations must implement.
+//
+// At runtime, either the Node.js or Web implementation is actually loaded. This works because the
+// package.json specifies "#node-web-compat" as a subpath import, with conditions pointing to the right file (for Node.js or Web)
 
 import { Json } from "./safe-json-parse.js";
 import {
@@ -11,8 +15,10 @@ import {
   JwsVerificationFunctionSync,
   JwsVerificationFunctionAsync,
 } from "./jwt-rsa";
-import { FetchError, NonRetryableFetchError } from "./error.js";
 
+/**
+ * Interface that the specific Node.js and Web implementations must implement
+ */
 export interface NodeWebCompat {
   transformJwkToKeyObjectSync: JwkToKeyObjectTransformerSync;
   transformJwkToKeyObjectAsync: JwkToKeyObjectTransformerAsync;
@@ -28,36 +34,4 @@ export interface NodeWebCompat {
     socketIdle?: number; // socket idle timeout (Only supported by Node.js runtime)
     response: number; // total round trip timeout
   };
-}
-
-/**
- * Sanity check a HTTPS response where we expect to get JSON data back
- *
- * @param uri the uri that was being requested
- * @param statusCode the HTTP status code, should be 200
- * @param contentType the value of the "Content-Type" header in the response, should start with "application/json"
- * @returns void - throws an error if the status code or content type aren't as expected
- */
-export function validateHttpsJsonResponse(
-  uri: string,
-  statusCode?: number,
-  contentType?: string
-): void {
-  if (statusCode === 429) {
-    throw new FetchError(uri, "Too many requests");
-  } else if (statusCode !== 200) {
-    throw new NonRetryableFetchError(
-      uri,
-      `Status code is ${statusCode}, expected 200`
-    );
-  }
-  if (
-    !contentType ||
-    !contentType.toLowerCase().startsWith("application/json")
-  ) {
-    throw new NonRetryableFetchError(
-      uri,
-      `Content-type is "${contentType}", expected "application/json"`
-    );
-  }
 }
